refactor(static-analysis): rename InitialStaticAnalysisRunner interface to InitialStaticAnalysis
Collapse the three-class hierarchy (interface InitialStaticAnalysisRunner +
concrete InitialStaticAnalysisProcessRunner + NullInitialStaticAnalysisRunner)
into two classes: InitialStaticAnalysis interface and InitialStaticAnalysisRunner
concrete implementation. InitialStaticAnalysisProcessRunner was a thin wrapper
with no reason for the extra indirection.
