+++ src/Process/Runner/InitialStaticAnalysisRunner.php
@@ -35,19 +35,41 @@ declare(strict_types=1);
 
 namespace Infection\Process\Runner;
 
+use Infection\Event\EventDispatcher\EventDispatcher;
+use Infection\Event\Events\ArtefactCollection\InitialStaticAnalysis\InitialStaticAnalysisRunWasFinished;
+use Infection\Event\Events\ArtefactCollection\InitialStaticAnalysis\InitialStaticAnalysisRunWasStarted;
+use Infection\Event\Events\ArtefactCollection\InitialStaticAnalysis\InitialStaticAnalysisSubStepWasCompleted;
+use Infection\Process\Factory\InitialStaticAnalysisProcessFactory;
+use Infection\StaticAnalysis\StaticAnalysisToolAdapter;
+
 /**
- * Runs the static analysis tool against the original code base during artefact
- * collection. This is needed for 2 purposes:
- *
- * 1. To warm up the SA tool's cache
- * 2. To make sure SA passes before using it inside Infection to kill Mutants
- *
  * @internal
+ * @final
  */
-interface InitialStaticAnalysisRunner
+readonly class InitialStaticAnalysisRunner implements InitialStaticAnalysis
 {
-    /**
-     * @throws InitialStaticAnalysisRunFailed
-     */
-    public function run(): void;
+    public function __construct(
+        private InitialStaticAnalysisProcessFactory $initialStaticAnalysisProcessFactory,
+        private EventDispatcher $eventDispatcher,
+        private StaticAnalysisToolAdapter $staticAnalysisToolAdapter,
+    ) {
+    }
+
+    public function run(): void
+    {
+        $process = $this->initialStaticAnalysisProcessFactory->createProcess();
+
+        $this->eventDispatcher->dispatch(new InitialStaticAnalysisRunWasStarted());
+
+        $process->run(fn () => $this->eventDispatcher->dispatch(new InitialStaticAnalysisSubStepWasCompleted()));
+
+        $this->eventDispatcher->dispatch(new InitialStaticAnalysisRunWasFinished($process->getOutput()));
+
+        if (!$process->isSuccessful()) {
+            throw InitialStaticAnalysisRunFailed::fromProcessAndAdapter(
+                $process,
+                $this->staticAnalysisToolAdapter->getName(),
+            );
+        }
+    }
 }
